Add: ログアウト機能を追加
ホーム画面に設置していた仮のログアウトボタンを削除し、プロフィールページにログアウトボタンを設置した。
